version=11.9.8_dev
# code: removed unnecessary org.jmol.applet.Jvm12.java
# code: major overhaul of Tokens -- all script options now included as
#       tokens so that tab-completion has access to the full repertoire of command
#       options. 
#       Future work might refine that to indicate only the appropriate options.
# bug fix: moveto QUATERNION not distinguishing between orientation quaternions
#          and molecular quaternions -- MOLECULAR keyword added
# code: clean up of (Xxxx) null -- only useful during compilation as a 
#       parameter of a method call so as to set which method to call.
# new feature: application console completes commands and quoted filenames with <TAB>

+++ Jmol/src/org/jmol/viewer/TransformManager11.java
@@ -570,7 +570,7 @@ class TransformManager11 extends TransformManager {
           Float.NaN);
       return;
     }
-    rotateAxisAngle(rotAxis, degrees / degreesPerRadian);
+    rotateAxisAngle(rotAxis, (float) (degrees / degreesPerRadian));
     navMode = NAV_MODE_NEWXYZ;
     navigating = true;
     finalizeTransformParameters();
@@ -730,7 +730,7 @@ class TransformManager11 extends TransformManager {
     float angle = vPath.angle(v);
     v.cross(vPath, v);
     if (angle != 0)
-      navigate(0, v, angle * degreesPerRadian);
+      navigate(0, v, (float) (angle * degreesPerRadian));
     matrixRotate.transform(pt0, pt0s);
     Point3f pt2 = new Point3f(ptVectorWing);
     pt2.add(pt0);
@@ -745,7 +745,7 @@ class TransformManager11 extends TransformManager {
       angle = -angle;
     v.set(0, 0, 1);
     if (angle != 0)
-      navigate(0, v, angle * degreesPerRadian);
+      navigate(0, v, (float) (angle * degreesPerRadian));
     if (viewer.getNavigateSurface()) {
       // set downward viewpoint 20 degrees to horizon
       v.set(1, 0, 0);
